refactor: centralize UI initialization
