+++ hud.js
@@ -16,16 +16,6 @@ function initHUD() {
   updateLeaderboard();
 }
 
-document.addEventListener('DOMContentLoaded', () => {
-=======
-  const leaderboardButton = document.getElementById('leaderboardButton');
-  const closeLeaderboard = document.getElementById('closeLeaderboard');
-  if (leaderboardButton)
-    leaderboardButton.addEventListener('click', showLeaderboard);
-  if (closeLeaderboard)
-    closeLeaderboard.addEventListener('click', hideLeaderboard);
-});
-
 function updateHUD({ score, highScore, lives, level }) {
   if (scoreEl) scoreEl.textContent = score;
   if (highScoreEl) highScoreEl.textContent = highScore;
