+++ index.html
@@ -7,8 +7,6 @@
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Press+Start+2P&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="styles.css">
-=======
-  <link rel="stylesheet" type="text/css" href="styles.css" />
 </head>
 <body>
   <div id="gameContainer">
@@ -58,22 +56,12 @@
     </div>
   </div>
 
-  <script src="starfield.js"></script>
-=======
   <script type="module">
     import Game from './game.js';
     import { initGameUI } from './ui.js';
 
     const game = new Game();
-    document.addEventListener('DOMContentLoaded', () => initGameUI(game));
+    initGameUI(game);
   </script>
-=======
-=======
-    const game = new Game();
-    document.addEventListener('DOMContentLoaded', () => initGameUI(game));
-  </script>
-=======
-  <script type="module" src="./game.js"></script>
 </body>
 </html>
-
